Enterprise-Database-Migration/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-db-migration"
version = "1.0.0"
description = "AI-Assisted Enterprise Database Migration System using LangGraph Multi-Agents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AI Migration Team"}
]
keywords = ["database", "migration", "ai", "langgraph", "postgresql", "mysql"]

dependencies = [
    # Core AI Frameworks
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-community>=0.3.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "psycopg2-binary>=2.9.9",
    
    # SQL Parsing & Transformation
    "sqlglot>=26.0.0",
    
    # UI
    "streamlit>=1.40.0",
    
    # CLI
    "rich>=13.9.0",
    "questionary>=2.1.0",
    "typer>=0.15.0",
    
    # Validation & Config
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "testcontainers>=4.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
db-migrate = "src.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
